New feature -- outputting to graphml
Instead of visualizing w/i python, I decided the better (read: easier) route is to output a graphml file to be used in an external tool instead (like Gephi). Everyone's map will be different, and this gives you the ability to customize it as you'd like.

This file reads the one output from the previous step (saved in the same folder), and creates nodes and edges based on it. You're given the option to include yourself as a node, but in my experience it just messes up the graph (for my purpose of identifying family groups, at least).