Multithreading
Experimental! Added multithreading. I ran into what appeared to be a rate limit on 23andMe's end, so it's limited to 4 threads. My test put it at ~20 minutes for 100 relatives, which is a lot faster than the non-multithreaded one. Once I clean up and refactor more, this will become the main branch